Added some more stuff:
- Changed player to a dictionary to be consistent with monsters being dictionaries.
Side effect is that all attacks can target anyone, as player and monster keys have the same name! This also allows for a the character to gain a spell that lets them use monster attacks against them and have it actually work in a fairly simple manner.

- Added the first attack in the game, scratch!
Now attacks have a framework that can be followed, taking in an attacker and a target as arguments. Spells will probably work the same way and while it's nice to have them separate, I would appreciate some discussion on whether at this point having a spells file is even necessary - we can use the attacks instead, since monsters could also cast spells.

- Game.py has a crappy proof of concept combat sample.
It shows how the scratch attack can be used by either a monster or the player, on either the monster or the player, and deal appropriate damage to the target's (current) health according to the attacker's stats.